--- bot.py
@@ -1,203 +1,0 @@
-from discord.ext import commands
-import discord
-from cogs.utils import context
-from cogs.utils.help import HelpCommand
-from cogs.utils.json import JSON
-from cogs.utils.config import Config
-from cogs.utils.embed import SimpleEmbed
-import datetime
-import json
-import logging
-import traceback
-import sys
-import aiohttp
-from collections import Counter, deque, defaultdict
-
-description = """
-another-discord-bot (dev branch)
-"""
-
-log = logging.getLogger(__name__)
-
-initial_extensions = (
-    'cogs.admin',
-    'cogs.mod',
-    'cogs.music',
-    'cogs.meta',
-    'cogs.reddit',
-    'cogs.events'
-)
-
-
-class ADB(commands.AutoShardedBot):
-    def __init__(self):
-        super().__init__(command_prefix='_a!', description=description,
-                         fetch_offline_members=False, heartbeat_timeout=150.0,
-                         help_command=HelpCommand())
-
-        self.session = aiohttp.ClientSession(loop=self.loop)
-
-        self._prev_events = deque(maxlen=10)
-
-        self.config = Config()
-        # shard_id: List[datetime.datetime]
-        # shows the last attempted IDENTIFYs and RESUMEs
-        self.resumes = defaultdict(list)
-        self.identifies = defaultdict(list)
-
-        # guild_id and user_id mapped to True
-        # these are users and guilds globally blacklisted
-        # from using the bot
-        self.blacklist = JSON('blacklist.json')
-
-        # in case of even further spam, add a cooldown mapping
-        # for people who excessively spam commands
-        self.spam_control = commands.CooldownMapping.from_cooldown(10, 12.0, commands.BucketType.user)
-
-        # A counter to auto-ban frequent spammers
-        # Triggering the rate limit 5 times in a row will auto-ban the user from the bot.
-        self._auto_spam_count = Counter()
-
-        for extension in initial_extensions:
-            try:
-                self.load_extension(extension)
-                log.info(f'Loaded {extension}')
-            except Exception as e:
-                log.error(f'Failed to load extension {extension} due to {e}.')
-                traceback.print_exc()
-
-    def _clear_gateway_data(self):
-        one_week_ago = datetime.datetime.utcnow() - datetime.timedelta(days=7)
-        for shard_id, dates in self.identifies.items():
-            to_remove = [index for index, dt in enumerate(dates) if dt < one_week_ago]
-            for index in reversed(to_remove):
-                del dates[index]
-
-        for shard_id, dates in self.resumes.items():
-            to_remove = [index for index, dt in enumerate(dates) if dt < one_week_ago]
-            for index in reversed(to_remove):
-                del dates[index]
-
-    async def on_socket_response(self, msg):
-        self._prev_events.append(msg)
-
-    async def before_identify_hook(self, shard_id, *, initial):
-        self._clear_gateway_data()
-        self.identifies[shard_id].append(datetime.datetime.utcnow())
-        await super().before_identify_hook(shard_id, initial=initial)
-
-    async def on_command_error(self, ctx, error):
-        if isinstance(error, commands.NoPrivateMessage):
-            await ctx.author.send('This command cannot be used in private messages.')
-        elif isinstance(error, commands.DisabledCommand):
-            await ctx.author.send('Sorry. This command is disabled and cannot be used.')
-        elif isinstance(error, commands.CommandInvokeError):
-            original = error.original
-            if not isinstance(original, discord.HTTPException):
-                print(f'In {ctx.command.qualified_name}:', file=sys.stderr)
-                traceback.print_tb(original.__traceback__)
-                print(f'{original.__class__.__name__}: {original}', file=sys.stderr)
-        elif isinstance(error, commands.ArgumentParsingError):
-            await ctx.send(error)
-
-    async def add_to_blacklist(self, object_id):
-        await self.blacklist.put(object_id, True)
-
-    async def remove_from_blacklist(self, object_id):
-        try:
-            await self.blacklist.remove(object_id)
-        except KeyError:
-            pass
-
-    async def on_ready(self):
-        if not hasattr(self, 'uptime'):
-            self.uptime = datetime.datetime.utcnow()
-
-        print(f'Ready: {self.user} (ID: {self.user.id})')
-        await self.change_presence(
-            activity=discord.Streaming(name=f'_a!help',
-                                       url='https://www.twitch.tv/commanderroot'))
-
-    async def on_shard_resumed(self, shard_id):
-        print(f'Shard ID {shard_id} has resumed...')
-        self.resumes[shard_id].append(datetime.datetime.utcnow())
-
-    @property
-    def stats_webhook(self):
-        wh_id= self.config.wh_id
-        wh_token = self.config.wh_token
-        hook = discord.Webhook.partial(id=wh_id, token=wh_token, adapter=discord.AsyncWebhookAdapter(self.session))
-        return hook
-
-    def log_spammer(self, ctx, message, retry_after, *, autoblock=False):
-        guild_name = getattr(ctx.guild, 'name', 'No Guild (DMs)')
-        guild_id = getattr(ctx.guild, 'id', None)
-        fmt = 'User %s (ID %s) in guild %r (ID %s) spamming, retry_after: %.2fs'
-        log.warning(fmt, message.author, message.author.id, guild_name, guild_id, retry_after)
-        if not autoblock:
-            return
-
-        wh = self.stats_webhook
-        e = SimpleEmbed(title='Auto-blocked Member')
-        e.add_field(name='Member', value=f'{message.author} (ID: {message.author.id})', inline=False)
-        e.add_field(name='Guild Info', value=f'{guild_name} (ID: {guild_id})', inline=False)
-        e.add_field(name='Channel Info', value=f'{message.channel} (ID: {message.channel.id}', inline=False)
-        e.timestamp = datetime.datetime.utcnow()
-        return wh.send(embed=e)
-
-    async def process_commands(self, message):
-        ctx = await self.get_context(message, cls=context.Context)
-
-        if ctx.command is None:
-            return
-
-        if ctx.author.id in self.blacklist:
-            return
-
-        if ctx.guild is not None and ctx.guild.id in self.blacklist:
-            return
-
-        bucket = self.spam_control.get_bucket(message)
-        current = message.created_at.replace(tzinfo=datetime.timezone.utc).timestamp()
-        retry_after = bucket.update_rate_limit(current)
-        author_id = message.author.id
-        if retry_after and author_id != self.owner_id:
-            self._auto_spam_count[author_id] += 1
-            if self._auto_spam_count[author_id] >= 5:
-                await self.add_to_blacklist(author_id)
-                del self._auto_spam_count[author_id]
-                await self.log_spammer(ctx, message, retry_after, autoblock=True)
-            else:
-                self.log_spammer(ctx, message, retry_after)
-            return
-        else:
-            self._auto_spam_count.pop(author_id, None)
-
-        await self.invoke(ctx)
-        log.info(f'Invoked command {message.content} by {author_id}')
-
-    async def on_message(self, message):
-        if message.author.bot:
-            return
-        await self.process_commands(message)
-
-    async def on_guild_join(self, guild):
-        if guild.id in self.blacklist:
-            await guild.leave()
-
-    async def close(self):
-        await super().close()
-        await self.session.close()
-
-    def run(self):
-        try:
-            super().run(self.config.login_token, reconnect=True)
-        finally:
-            with open('logs/prev_events.log', 'w', encoding='utf-8') as fp:
-                for data in self._prev_events:
-                    try:
-                        x = json.dumps(data, ensure_ascii=True, indent=4)
-                    except:
-                        fp.write(f'{data}\n')
-                    else:
-                        fp.write(f'{x}\n')
