+++ config/example_options.ini
@@ -10,9 +10,7 @@ DevIDs =
 
 [Bot]
 
-WebhookID =
-
-WebhookToken =
+Prefix = _a!
 
 EnableMSGLogging = True
