+++ core/cogs/events/events.py
@@ -1,7 +1,7 @@
 import discord
 from discord.ext import commands
 import logging
-from .utils import checks
+from ...utils import checks
 
 log = logging.getLogger(__name__)
