+++ core/utils/checks.py
@@ -4,6 +4,8 @@ config = Config()
 
 # those checks are written without security or extended functionality in mind
 # they don't take channel overrides into account
+# needs a lot of tuning still
+# TODO user overridable admin and mod roles
 
 async def check_permissions(ctx, perms, *, check=all):
     is_owner = await ctx.bot.is_owner(ctx.author)
