+++ run.sh
@@ -17,14 +17,9 @@ sleep 5
 
 echo "Checking for python venv"
 
-venvdir="./adb/"
- 
-url="some_url/file.tar.gz"
-file="${url##*/}"
- 
-[ ! -d "$venvdir" ] && python -m venv "$dldir" && echo "Venv created... attempting to launch bot"
+venvdir="./venv/"
  
 echo "Launching bot in screen adb"
 
-screen -dmLS adb ./adb/bin/python launcher.py
+screen -dmLS adb ./venv/bin/python run.py
